% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-activate_wearable_dataset.R
\name{activate_wearable_dataset}
\alias{activate_wearable_dataset}
\title{Activate Specific Dataset in wearabledataset Package}
\usage{
activate_wearable_dataset(
  .data,
  what = c("sample_info", "variable_info", "expression_data")
)
}
\arguments{
\item{.data}{An object of class \code{wearable_dataset}.}

\item{what}{A character vector specifying which dataset to activate.
Can be one of "sample_info", "variable_info", or "expression_data".}
}
\value{
A modified version of the \code{wearable_dataset} object with the "activated" slot set.
}
\description{
This function sets the "activated" slot of a given \code{wearable_dataset} object
from the wearabledataset package to the specified dataset type.
}
\examples{
\dontrun{
dataset <- your_dataset_function() # replace with how you get a wearable_dataset object
activate_wearable_dataset(dataset, what = "sample_info")
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
