% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-check_wearable_dataset.R
\name{check_wearable_dataset}
\alias{check_wearable_dataset}
\title{Validate Components of a \code{wearable_dataset} Structure}
\usage{
check_wearable_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{A data.frame representing the expression data.}

\item{sample_info}{A data.frame representing the sample information.}

\item{variable_info}{A data.frame representing the variable information.}

\item{sample_info_note}{(Optional) A data.frame representing additional notes on the sample information.}

\item{variable_info_note}{(Optional) A data.frame representing additional notes on the variable information.}
}
\value{
A character string detailing the status of the validation. If there are any issues,
it returns a string starting with "error:". If all checks pass, it returns "all good.".
}
\description{
This function checks the individual components that make up the structure of
a \code{wearable_dataset} object from the wearabledataset package. It ensures that the components
meet specific criteria and standards.
}
\details{
The function checks each parameter's type and relevant components:
\itemize{
\item Ensures \code{expression_data}, \code{sample_info}, and \code{variable_info} are provided.
\item Validates the structure and required columns for \code{sample_info}, \code{variable_info}, \code{sample_info_note},
and \code{variable_info_note}.
\item Checks consistency and correspondence between columns, row names, and IDs among the provided datasets.
}
}
\examples{
\dontrun{
expression_data_sample <- data.frame() # replace with sample expression data
sample_info_sample <- data.frame() # replace with sample sample_info data
variable_info_sample <- data.frame() # replace with sample variable_info data

check_wearable_dataset(expression_data_sample, sample_info_sample, variable_info_sample)
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
