% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-check_wearable_dataset_class.R
\docType{methods}
\name{check_wearable_dataset_class}
\alias{check_wearable_dataset_class}
\title{check_wearable_dataset_class}
\usage{
check_wearable_dataset_class(object)
}
\arguments{
\item{object}{wearable_dataset class.}
}
\value{
Notice of data checking.
}
\description{
Check wearable_dataset class object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
