% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-all_classes.R
\docType{methods}
\name{create_wearable_dataset}
\alias{create_wearable_dataset}
\title{create_wearable_dataset}
\usage{
create_wearable_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{expression_data.}

\item{sample_info}{Sample information name.}

\item{variable_info}{Variable information.
Columns are samples and rows are variables.}

\item{sample_info_note}{Sample information name.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}

\item{variable_info_note}{Sample information name.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}
}
\value{
A wearable_dataset-class object.
}
\description{
Create the wearable_dataset object.
}
\examples{
data("expression_data", package = "wearabledataset")
data("variable_info", package = "wearabledataset")
data("sample_info", package = "wearabledataset")
object <-
create_wearable_dataset(expression_data = expression_data,
sample_info = sample_info,
variable_info = variable_info)
object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
