% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-graphs.R
\name{draw_time_plot}
\alias{draw_time_plot}
\alias{draw_time_plot.numeric}
\alias{draw_time_plot.wearable_dataset}
\alias{draw_time_plot.default}
\title{Draw time serial plot}
\usage{
draw_time_plot(
  x,
  accurate_time,
  variable_index,
  color = "blue",
  y_axis_name = "Value",
  sun_rise_time = "6:00:00",
  sun_set_time = "18:00:00",
  time_gap = 12,
  add_point = FALSE,
  facet = FALSE,
  facet_function = c("facet_grid", "facet_wrap"),
  facet_grid_by = c("rows", "cols"),
  facet_wrap_nrow,
  facet_wrap_ncol,
  ...
)

\method{draw_time_plot}{numeric}(
  x,
  accurate_time,
  variable_index,
  color = "blue",
  y_axis_name = "Value",
  sun_rise_time = "6:00:00",
  sun_set_time = "18:00:00",
  time_gap = 12,
  add_point = FALSE,
  facet = FALSE,
  facet_function = c("facet_grid", "facet_wrap"),
  facet_grid_by = c("rows", "cols"),
  facet_wrap_nrow,
  facet_wrap_ncol,
  ...
)

\method{draw_time_plot}{wearable_dataset}(
  x,
  accurate_time,
  variable_index,
  color = "blue",
  y_axis_name = "Value",
  sun_rise_time = "6:00:00",
  sun_set_time = "18:00:00",
  time_gap = 12,
  add_point = FALSE,
  facet = FALSE,
  facet_function = c("facet_grid", "facet_wrap"),
  facet_grid_by = c("rows", "cols"),
  facet_wrap_nrow,
  facet_wrap_ncol,
  ...
)

\method{draw_time_plot}{default}(x, ...)
}
\arguments{
\item{x}{A numeric vector or wearable_dataset object.}

\item{accurate_time}{Accurate time}

\item{variable_index}{index of variable}

\item{color}{color}

\item{y_axis_name}{y_axis_name}

\item{sun_rise_time}{should be 24 hour format, default "6:00:00"}

\item{sun_set_time}{should be 24 hour format, default "18:00:00"}

\item{time_gap}{time gap for x axis text.}

\item{add_point}{add point or not.}

\item{facet}{facet or not.}

\item{facet_function}{facet function, facet_grid or facet_wrap}

\item{facet_grid_by}{facet_grid, rows or cols}

\item{facet_wrap_nrow}{nrow for facet_wrap}

\item{facet_wrap_ncol}{ncol for facet_wrap,}

\item{...}{othter arguments}
}
\value{
A ggplot2 object.
}
\description{
Draw time series plot
}
\examples{
data("expression_data", package = "wearabledataset")
data("sample_info", package = "wearabledataset")
data("variable_info", package = "wearabledataset")
object <-
create_wearable_dataset(expression_data, sample_info, variable_info)
object <- object \%>\% mutate2sample(what = "date")
unique(object@sample_info$date)
object <-
object \%>\% activate_wearable_dataset(what = "sample_info") \%>\%
filter(date \%in\% c("2022-08-08", "2022-08-09"))
expression_data <- extract_expression_data(object)
sample_info <- extract_sample_info(object)
x = as.numeric(expression_data[1,])
accurate_time = sample_info$accurate_time
draw_time_plot(x = x, accurate_time = accurate_time)
draw_time_plot(x = object, variable_index = 1,
facet = TRUE,
facet_function = "facet_wrap")
draw_time_plot(x = object,
               variable_index = 1,
               facet = TRUE,
               facet_function = "facet_grid",
               facet_grid_by = "rows", time_gap = 2)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
