% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-wearabledataset_logo.R
\docType{methods}
\name{get_wearabledataset_version}
\alias{get_wearabledataset_version}
\title{get_wearabledataset_version}
\usage{
get_wearabledataset_version()
}
\value{
version
}
\description{
Get wearabledataset package version
}
\examples{
get_wearabledataset_version()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
