% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-methods_mutate2sample.R
\name{mutate2sample}
\alias{mutate2sample}
\alias{mutate2sample.wearable_dataset}
\title{Mutate Sample-based Attributes to a \code{wearable_dataset} Object}
\usage{
mutate2sample(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number", "na_prop",
    "date", "week", "time"),
  according_to_variables = "all",
  ...
)

\method{mutate2sample}{wearable_dataset}(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity", "na_number", "na_prop",
    "date", "week", "time"),
  according_to_variables = "all",
  ...
)
}
\arguments{
\item{object}{A \code{wearable_dataset} object to be mutated.}

\item{what}{A character vector specifying which attributes to compute or mutate. Possible values include
"sum_intensity", "mean_intensity", "median_intensity", "na_number", "na_prop", "date", "week", and "time".}

\item{according_to_variables}{Specifies which variables to consider when performing the mutation.
Defaults to "all".}

\item{...}{Additional parameters to be passed on.}
}
\value{
An object with mutated or added sample-based attributes.
}
\description{
This generic function mutates or adds specific attributes to samples in the provided \code{wearable_dataset} object.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
