% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-all_classes.R
\docType{class}
\name{wearable_dataset-class}
\alias{wearable_dataset-class}
\title{wearable_dataset class}
\description{
setClassUnion("treedata_or_null", c("treedata", "NULL"))
}
\details{
setClassUnion("xstringset_or_null", c("XStringSet", "NULL"))
}
\section{Slots}{

\describe{
\item{\code{...}}{Other slots from mass_dataset class object in massdataset pacakge}
}}

