% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-wearabledataset_logo.R
\docType{methods}
\name{wearabledataset_logo}
\alias{wearabledataset_logo}
\title{wearabledataset_logo}
\usage{
wearabledataset_logo()
}
\value{
logo
}
\description{
Get the detailed information of wearabledataset package.
}
\examples{
wearabledataset_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
