% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-utils.R
\name{wearabledataset_packages}
\alias{wearabledataset_packages}
\title{List all packages in the wearabledataset}
\usage{
wearabledataset_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include wearabledataset in the list?}
}
\value{
wearabledataset packages
}
\description{
List all packages in the wearabledataset
}
\examples{
wearabledataset_packages()
}
