% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-check_wearable_dataset.R
\docType{methods}
\name{check_wearable_dataset}
\alias{check_wearable_dataset}
\title{check_wearable_dataset}
\usage{
check_wearable_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{expression_data.}

\item{sample_info}{sample_info.}

\item{variable_info}{variable_info.}

\item{sample_info_note}{sample_info_note.}

\item{variable_info_note}{variable_info_note.}
}
\value{
Notice of data checking.
}
\description{
Check data format for wearable_dataset class object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
