% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-check_wearable_dataset_class.R
\name{check_wearable_dataset_class}
\alias{check_wearable_dataset_class}
\title{Validate \code{wearable_dataset} Object Structure}
\usage{
check_wearable_dataset_class(object)
}
\arguments{
\item{object}{An object of class \code{wearable_dataset} to be validated.}
}
\value{
A logical value \code{TRUE} if the object passes all checks. If there are any issues,
it returns a character vector detailing the errors found.
}
\description{
This function checks the integrity and structure of a given \code{wearable_dataset} object
from the wearabledataset package to ensure that it meets specific criteria and standards.
}
\details{
The function checks:
\itemize{
\item \code{variable_info}, \code{sample_info}, \code{sample_info_note}, and \code{variable_info_note} for the presence
of required columns and their uniqueness.
\item \code{accurate_time} in \code{sample_info} for being of class \code{POSIXct}.
\item Consistency between the number of rows and columns in \code{expression_data} and other slots.
}
}
\examples{
\dontrun{
dataset <- your_dataset_function() # replace with how you obtain a wearable_dataset object
check_wearable_dataset_class(dataset)
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
