% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-all_classes.R
\docType{methods}
\name{create_wearable_dataset}
\alias{create_wearable_dataset}
\title{create_wearable_dataset}
\usage{
create_wearable_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{MS1 peak table name.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}

\item{sample_info}{Sample information name.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}

\item{variable_info}{MS1 peak table name.
Columns are samples and rows are variables.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}

\item{sample_info_note}{Sample information name.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}

\item{variable_info_note}{Sample information name.
\url{https://tidywearable.github.io/wearabledataset/articles/data_import_and_export.html}}
}
\value{
A wearable_dataset-class object.
}
\description{
Create the wearable_dataset object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
