% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-methods_mutate2variable.R
\name{mutate2variable}
\alias{mutate2variable}
\alias{mutate2variable.wearable_dataset}
\title{Mutate Variable-based Attributes}
\usage{
mutate2variable(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number", "na_prop"),
  according_to_samples = "all",
  ...
)

\method{mutate2variable}{wearable_dataset}(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number", "na_prop"),
  according_to_samples = "all",
  ...
)
}
\arguments{
\item{object}{An object to be mutated. The specific class-based method to use will depend on the class of this object.}

\item{what}{A character vector specifying which attributes to compute or mutate. Possible values include
"mean_intensity", "median_intensity", "sum_intensity", "na_number", and "na_prop".}

\item{according_to_samples}{Specifies which samples to consider when performing the mutation.
Defaults to "all".}

\item{...}{Further arguments to be passed to specific methods.}
}
\value{
An object with mutated or added variable-based attributes.
}
\description{
This generic function mutates or adds specific attributes to variables in the provided object.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
